pyproject.toml
[tool.poetry]
name = "mare-demo"
version = "0.1.0"
description = ""
authors = ["Takuma Fukuda <takuma.fukuda@mare-key.ai>"]
readme = "README.md"
packages = [{include = "configs"}, {include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
openai = "^1.33.0"
notebook = "^7.2.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
scikit-learn = "^1.5.0"
numpy = "^1.26.4"
plotly = "^5.22.0"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
langchain-openai = "^0.1.9"
langchain-experimental = "^0.0.61"
neo4j = "^5.21.0"
wikipedia = "^1.4.0"
tiktoken = "^0.7.0"
yfiles-jupyter-graphs = "^1.7.2"
streamlit = "^1.36.0"
python-docx = "^1.1.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
